fix: update input array to double

+++ src/main/java/br/edu/ifsp/testing/practice/TriangleUtils.java
@@ -6,7 +6,7 @@ public class TriangleUtils {
 
     public record Statistics(int nonTriangle, int equilateral, int isosceles, int scalene) {}
 
-    public Statistics computeTypeStatistics(int[][] data) {
+    public Statistics computeTypeStatistics(double[][] data) {
         Objects.requireNonNull(data);
 
         int nonTriangle = 0;
@@ -15,12 +15,12 @@ public class TriangleUtils {
         int scalene = 0;
 
         for (int i = 0; i < data.length; i++) {
-            int[] datum = data[i];
+            double[] datum = data[i];
             if (datum.length != 6) throw new IllegalArgumentException();
 
-            int x1 = datum[0]; int y1 = datum[1];
-            int x2 = datum[2]; int y2 = datum[3];
-            int x3 = datum[4]; int y3 = datum[5];
+            double x1 = datum[0]; double y1 = datum[1];
+            double x2 = datum[2]; double y2 = datum[3];
+            double x3 = datum[4]; double y3 = datum[5];
 
             double a = distance(x1, y1, x2, y2);
             double b = distance(x2, y2, x3, y3);
@@ -34,9 +34,9 @@ public class TriangleUtils {
         return new Statistics(nonTriangle, equilateral, isosceles, scalene);
     }
 
-    private double distance(int x1, int y1, int x2, int y2) {
-        int dx = x1 - x2;
-        int dy = y1 - y2;
+    private double distance(double x1, double y1, double x2, double y2) {
+        double dx = x1 - x2;
+        double dy = y1 - y2;
         final double euclideanDistance = Math.sqrt(dx * dx + dy * dy);
         return Math.round(euclideanDistance * 1000.0) / 1000.0;
     }
